+++ framework/model-test/src/main/java/is/codion/framework/model/test/AbstractEntityApplicationModelTest.java
@@ -21,9 +21,9 @@ package is.codion.framework.model.test;
 import is.codion.common.utilities.user.User;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
-import is.codion.framework.model.DefaultEntityApplicationModel;
-import is.codion.framework.model.DefaultEntityEditModel;
-import is.codion.framework.model.DefaultEntityModel;
+import is.codion.framework.model.AbstractEntityApplicationModel;
+import is.codion.framework.model.AbstractEntityEditModel;
+import is.codion.framework.model.AbstractEntityModel;
 import is.codion.framework.model.EntityApplicationModel;
 import is.codion.framework.model.EntityEditModel;
 import is.codion.framework.model.EntityEditor;
@@ -33,6 +33,8 @@ import is.codion.framework.model.test.TestDomain.Department;
 import is.codion.framework.model.test.TestDomain.Employee;
 
 import org.junit.jupiter.api.Test;
+
+import java.util.Collection;
 
 import static java.util.Collections.emptyList;
 import static java.util.Collections.singleton;
@@ -45,8 +47,8 @@ import static org.junit.jupiter.api.Assertions.*;
  * @param <T> the {@link EntityTableModel} type
  * @param <R> the {@link EntityEditor} type
  */
-public abstract class AbstractEntityApplicationModelTest<M extends DefaultEntityModel<M, E, T, R>,
-				E extends DefaultEntityEditModel<R>, T extends EntityTableModel<E, R>, R extends EntityEditor<R>> {
+public abstract class AbstractEntityApplicationModelTest<M extends AbstractEntityModel<M, E, T, R>,
+				E extends AbstractEntityEditModel<R>, T extends EntityTableModel<E, R>, R extends EntityEditor<R>> {
 
 	private static final User UNIT_TEST_USER =
 					User.parse(System.getProperty("codion.test.user", "scott:tiger"));
@@ -65,7 +67,7 @@ public abstract class AbstractEntityApplicationModelTest<M extends DefaultEntity
 	@Test
 	public void applicationModel() {
 		M deptModel = createDepartmentModel();
-		EntityApplicationModel<M, E, T, R> model = new DefaultEntityApplicationModel<>(connectionProvider, singleton(deptModel));
+		EntityApplicationModel<M, E, T, R> model = new TestEntityApplicationModel<>(connectionProvider, singleton(deptModel));
 		assertNotNull(model.models().get(Department.TYPE));
 		assertEquals(1, model.models().get().size());
 		assertEquals(UNIT_TEST_USER, model.user());
@@ -75,33 +77,33 @@ public abstract class AbstractEntityApplicationModelTest<M extends DefaultEntity
 
 	@Test
 	public void constructorNullConnectionProvider() {
-		assertThrows(NullPointerException.class, () -> new DefaultEntityApplicationModel<>(null, emptyList()));
+		assertThrows(NullPointerException.class, () -> new TestEntityApplicationModel<>(null, emptyList()));
 	}
 
 	@Test
 	public void entityModelByEntityTypeNotFound() {
-		EntityApplicationModel<M, E, T, R> model = new DefaultEntityApplicationModel<>(connectionProvider, emptyList());
+		EntityApplicationModel<M, E, T, R> model = new TestEntityApplicationModel<>(connectionProvider, emptyList());
 		assertThrows(IllegalArgumentException.class, () -> model.models().get(Department.TYPE));
 	}
 
 	@Test
 	public void entityModelByEntityType() {
 		M departmentModel = createDepartmentModel();
-		EntityApplicationModel<M, E, T, R> model = new DefaultEntityApplicationModel<>(connectionProvider, singleton(departmentModel));
+		EntityApplicationModel<M, E, T, R> model = new TestEntityApplicationModel<>(connectionProvider, singleton(departmentModel));
 		assertEquals(departmentModel, model.models().get(Department.TYPE));
 	}
 
 	@Test
 	public void entityModelByClass() {
 		M departmentModel = createDepartmentModel();
-		EntityApplicationModel<M, E, T, R> model = new DefaultEntityApplicationModel<>(connectionProvider, singleton(departmentModel));
+		EntityApplicationModel<M, E, T, R> model = new TestEntityApplicationModel<>(connectionProvider, singleton(departmentModel));
 		assertEquals(departmentModel, model.models().get((Class<? extends M>) departmentModel.getClass()));
 	}
 
 	@Test
 	public void containsEntityModel() {
 		M departmentModel = createDepartmentModel();
-		EntityApplicationModel<M, E, T, R> model = new DefaultEntityApplicationModel<>(connectionProvider, singleton(departmentModel));
+		EntityApplicationModel<M, E, T, R> model = new TestEntityApplicationModel<>(connectionProvider, singleton(departmentModel));
 
 		assertTrue(model.models().contains(Department.TYPE));
 		assertTrue(model.models().contains((Class<? extends M>) departmentModel.getClass()));
@@ -121,4 +123,16 @@ public abstract class AbstractEntityApplicationModelTest<M extends DefaultEntity
 	 * @see Department#TYPE
 	 */
 	protected abstract M createDepartmentModel();
+
+	private static final class TestEntityApplicationModel<
+					M extends EntityModel<M, E, T, R>,
+					E extends EntityEditModel<R>,
+					T extends EntityTableModel<E, R>,
+					R extends EntityEditor<R>>
+					extends AbstractEntityApplicationModel<M, E, T, R> {
+
+		private TestEntityApplicationModel(EntityConnectionProvider connectionProvider, Collection<? extends M> entityModels) {
+			super(connectionProvider, entityModels);
+		}
+	}
 }