+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.79
 ### is.codion.framework.model
 - AbstractEntityEditor.execute() and supersede() protected again.
+- DefaultEntityModel, DefaultEntityEditModel and DefaultEntityApplicationModel renamed AbstractEntityModel, AbstractEntityEditModel and AbstractEntityApplicationModel, now abstract, consistent with the other Abstract* extension bases.
 
 ## 0.18.78
 ### is.codion.common.reactive